refactor: 코드 이동 및 정리
