--- src/main/java/roomescape/dto/WaitingResponse.java
@@ -1,9 +1,0 @@
-package roomescape.dto;
-
-import roomescape.domain.Reservation;
-
-public record WaitingResponse(long id, String name, int order) {
-    public static WaitingResponse from(Reservation reservation, int order) {
-        return new WaitingResponse(reservation.getId(), reservation.getName(), order);
-    }
-}
